Multi variable optimization code completed
The file optimization_project_02.py, which optimizes the dimensions of beams and columns of a 3d frame structure is completed. The variables involved in this optimization are four: deepth and width of the beams and deepth and width of columns. All other modules were updated in order to perform the optimization without losing the ability to solve traditional problems which does not involve optimization.